Descifrado

+++ Lab2Cifrados/Models/RSA.cs
@@ -111,8 +111,9 @@ namespace Lab2Cifrados.Models
 
         public int CifYDescifNumero(int NCifrar, int llave, int n)
         {
-            /*ulong potencia = Convert.ToUInt64(Math.Pow(Convert.ToUInt64(NCifrar), Convert.ToUInt64(llave)));
-            var cifrado = Convert.ToInt32(potencia % Convert.ToUInt64(n));*/
+            
+            /*double potencia = Math.Pow(NCifrar, llave);
+            var cifrado = Convert.ToInt32(potencia % Convert.ToInt32(n));*/
             int valor = 1;
             for (int i = 0; i < llave; i++)
             {
